2.1pre37
{
    "filename": "links-2.1pre37.tar.gz",
    "md5": "f26f55b45e98d0cba47acf966014e4de",
    "sha1": "fbf7df67a5b06ccf06901664687780a4bc6a7963",
    "sha256": "8552ea061f086e55e98391737927af5a8982da7f36732436eeee3e3f5753cf3a"
}

+++ bfu.c
@@ -7,11 +7,11 @@
 
 int menu_font_size=G_BFU_DEFAULT_FONT_SIZE;
 
-unsigned long G_BFU_FG_COLOR;
-unsigned long G_BFU_BG_COLOR;
-unsigned long G_SCROLL_BAR_AREA_COLOR;
-unsigned long G_SCROLL_BAR_BAR_COLOR;
-unsigned long G_SCROLL_BAR_FRAME_COLOR;
+unsigned G_BFU_FG_COLOR;
+unsigned G_BFU_BG_COLOR;
+unsigned G_SCROLL_BAR_AREA_COLOR;
+unsigned G_SCROLL_BAR_BAR_COLOR;
+unsigned G_SCROLL_BAR_FRAME_COLOR;
 
 /* prototypes */
 int is_in_str(unsigned char *, int);
